copilot_agent/requirements.txt
# Copilot Agent - Dependencias
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
